feat: Corrected initial implementation of Retrofit
The initial implementation of getting the jokes from ICNDB
wasn't working correctly. This is because the root of the JSON
response wasn't taken into account initially. Due to that, the
HTTP response was always coming back as null. The issue has been
corrected by creating a new class called ICNDBResponse that
mimicks the root of the JSON response. The code will be cleaned
up more in future commits.