feat: Took out all Retrofit code
The previous Retrofit implementation was done mostly for fun. It
was also done to get better practice with Retrofit because it is
a very powerful and popular library. All of the previous work
was set aside into a separate branch called
RetrofitImplementation. The ChuckNorris class has been cleaned up
by removing all of the previous code. Classes that dealt with
Retrofit have since been deleted. This has all been done because
the Retrofit implementation doesn't align with the requirements
of the Udacity project.