feat: Adjusted EndpointsAsyncTask to be its own class
Adjusted the EndpointsAsyncTask to be a class of its own. This
was driven per a recommendation given by a mentor at Udacity
to have the EndpointsAsyncTask have an interface and for the
activity to implement the interface. This approach keeps things
modular. The discussion that inspired this approach can be
found here:
https://discussions.udacity.com/t/modifying-gce-starter-code-and-creating-async-task/473168/9