feat: Adjusted both free and paid variants
Made adjustments to both the paid and free build variants so that
the free variant still showed ads but the paid variant no longer
did. This involved adjustments to both the main manifest file and
the manifest of the free variant. Also, the MainActivityFragment
had to be adjusted in the free variant so that ads were still
being shown. Finally, the dependency of the Google ads services
was adjusted so that the free variant was the only one that
dependended on it.