feat: Started working on interstitial ads
I started adding interstital ads to the free flavor of the app.
However, I wanted the ads to be shown smoothly with the joke
being shown immediately after the ad is shown. I am currently
attempting to get this to work using CountdownLatch, but the app
becomes unresponsive with this implementation.