feat: Started working on ProgressBar
Started working on adding a ProgressBar for the paid flavor,
although the implementation would be the same for the free flavor
as well. For some reason, the ProgressBar is being shown before
the AsyncTask is executed, which is when the joke button is
clicked. I will need to look into this further.